+++ packages/crypto-signature-bls12-381/source/service-provider.ts
@@ -1,0 +1,13 @@
+import { Identifiers } from "@mainsail/constants";
+import { injectable } from "@mainsail/container";
+import { Providers } from "@mainsail/kernel";
+
+import { Signature } from "./signature.js";
+
+@injectable()
+export class ServiceProvider extends Providers.ServiceProvider {
+	public async register(): Promise<void> {
+		this.app.bind(Identifiers.Cryptography.Signature.Size).toConstantValue(96);
+		this.app.bind(Identifiers.Cryptography.Signature.Instance).to(Signature).inSingletonScope();
+	}
+}
